More coverity fixes

+++ reflector/StorageReflectSession.cpp
@@ -891,13 +891,13 @@ MessageRef StorageReflectSession :: GetEffectiveParameters() const
    (void) resultMessage()->AddString(PR_NAME_SERVER_VERSION, MUSCLE_VERSION_STRING);
 
    (void) resultMessage()->RemoveName(PR_NAME_SERVER_MEM_AVAILABLE);
-   (void) resultMessage()->AddInt64(PR_NAME_SERVER_MEM_AVAILABLE, GetNumAvailableBytes());
+   (void) resultMessage()->AddInt64(PR_NAME_SERVER_MEM_AVAILABLE, (int64) GetNumAvailableBytes());
 
    (void) resultMessage()->RemoveName(PR_NAME_SERVER_MEM_USED);
-   (void) resultMessage()->AddInt64(PR_NAME_SERVER_MEM_USED, GetNumUsedBytes());
+   (void) resultMessage()->AddInt64(PR_NAME_SERVER_MEM_USED, (int64) GetNumUsedBytes());
 
    (void) resultMessage()->RemoveName(PR_NAME_SERVER_MEM_MAX);
-   (void) resultMessage()->AddInt64(PR_NAME_SERVER_MEM_MAX, GetMaxNumBytes());
+   (void) resultMessage()->AddInt64(PR_NAME_SERVER_MEM_MAX, (int64) GetMaxNumBytes());
 
    const uint64 now = GetRunTime64();
